* device/include/pic16/limits.h, * device/lib/_divsint.c, * device/lib/_divslong.c, * device/lib/_divuint.c, * device/lib/_divulong.c, * device/lib/_memset.c, * device/lib/_modsint.c, * device/lib/_modslong.c, * device/lib/_moduint.c, * device/lib/_modulong.c, * device/lib/_mulint.c, * device/lib/_mullong.c, * device/lib/hc08/_divuint.c, * device/lib/hc08/_divulong.c, * device/lib/pic16/libm/asincosf.c, * device/lib/pic16/libm/atanf.c, * device/lib/pic16/libm/logf.c, * device/lib/s08/_divuint.c, * device/lib/s08/_divulong.c: replaced deprecated predefined macros * device/lib/pic16/libio/i2c/i2copen.c: idem to fix bug 2336 * src/SDCCast.c (ast_print), * src/SDCCicode.c (dbuf_printOperand): fixed warnings * support/regression/tests/bug-2313.c: fixed warning
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@9151 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCicode.c
@@ -168,7 +168,7 @@ dbuf_printOperand (operand * op, struct dbuf_s *dbuf)
       else if (IS_FIXED16X16 (opetype))
         dbuf_printf (dbuf, "%g {", doubleFromFixed16x16 (SPEC_CVAL (opetype).v_fixed16x16));
       else if (IS_LONGLONG (opetype))
-        dbuf_printf (dbuf, "0x%llx {", SPEC_CVAL (OP_VALUE (op)->etype).v_ulonglong);
+        dbuf_printf (dbuf, "0x%llx {", (unsigned long long) SPEC_CVAL (OP_VALUE (op)->etype).v_ulonglong);
       else
         dbuf_printf (dbuf, "0x%x {", (unsigned int) ulFromVal (OP_VALUE (op)));
       dbuf_printTypeChain (operandType (op), dbuf);
